results_writer/requirements.txt
confluent-kafka==2.2.0
psycopg2-binary==2.9.6
pandas==2.0.3
python-dotenv==1.0.0
kafka-python==2.0.2